requirements.txt
pandas
numpy
matplotlib
seaborn
kaggle
python-dotenv
jupyter
